refactor(clients): flatten hierarchy — BaseClient as root, delete ClientInterface
Replaces the 3-level hierarchy (ClientInterface → AuthenticatedClient
→ BaseSQLClient/BaseClient) with a flat one:

  BaseClient (root)
      ├── BaseSQLClient     — SQLAlchemy
      ├── BaseHTTPClient    — httpx (renamed from BaseClient)
      └── AzureClient       — Azure SDK

- ClientInterface ABC deleted (was just load + close stubs)
- AuthenticatedClient merged into BaseClient in clients/__init__.py
- Current BaseClient renamed to BaseHTTPClient in clients/base.py
- AzureClient now extends BaseClient (gains AUTH_STRATEGIES for free)
- All 93 connectors will extend one of these three
