+++ tests/unit/clients/test_base_client.py
@@ -6,7 +6,7 @@ import pytest
 from httpx import Headers
 from hypothesis import HealthCheck, given, settings
 
-from application_sdk.clients.base import BaseClient
+from application_sdk.clients.base import BaseHTTPClient as BaseClient
 from application_sdk.test_utils.hypothesis.strategies.clients.sql import (
     sql_credentials_strategy,
 )
